Change UserApiStatus.kt to NetworkUtil.kt and add it to utils package
