+++ app/src/main/java/com/example/githubdemo/utils/NetworkUtil.kt
@@ -1,0 +1,31 @@
+package com.example.githubdemo.utils
+
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities.*
+import android.os.Build
+
+
+class NetworkUtil(private val context: Context) {
+    var lastPage = false
+    fun hasInternetConnection(): Boolean {
+        val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as
+                ConnectivityManager
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            val activeNetwork = connectivityManager.activeNetwork ?: return false
+            val capabilities = connectivityManager.getNetworkCapabilities(activeNetwork)
+                ?: return false
+            return when {
+                capabilities.hasTransport(TRANSPORT_WIFI) -> true
+                capabilities.hasTransport(TRANSPORT_CELLULAR) -> true
+                capabilities.hasTransport(TRANSPORT_ETHERNET) -> true
+                else -> false
+            }
+        }
+        return false
+    }
+
+    fun isLastPage(isLastPage: Boolean) {
+        lastPage = isLastPage
+    }
+}
